refactor(selector4nix): rename modules and `Nar` to `NarInfo`
