+++ src/application/nar/actor/mod.rs
@@ -4,6 +4,6 @@ pub use runner::{NarActor, NarRequest, ResolveNarInfoResponse};
 
 use selector4nix_actor::registry::Registry;
 
-use crate::domain::nar::model::StorePathHash;
+use crate::domain::nar_info::model::StorePathHash;
 
 pub type NarActorRegistry = Registry<StorePathHash, NarActor>;
