+++ src/domain/mod.rs
@@ -1,2 +1,2 @@
-pub mod nar;
+pub mod nar_info;
 pub mod substituter;
