+++ src/domain/nar_info/index/mod.rs
