+++ src/domain/nar_info/model/mod.rs
@@ -1,9 +1,9 @@
-mod nar;
 mod nar_file_name;
+mod nar_info;
 mod nar_info_data;
 mod store_path_hash;
 
-pub use nar::{Nar, NarInfoResolution, NarUrlRewriteOption};
 pub use nar_file_name::{NarFileName, TryNewNarFileNameError};
+pub use nar_info::{NarInfo, NarInfoResolution, NarUrlRewriteOption};
 pub use nar_info_data::{NarInfoData, TryNewNarInfoData};
 pub use store_path_hash::{StorePathHash, TryNewStorePathHashError};
