+++ src/domain/nar_info/model/store_path_hash.rs
