+++ src/domain/nar_info/service/util.rs
