+++ src/infrastructure/provider/nar_stream_provider.rs
@@ -7,8 +7,8 @@ use reqwest::{Client, Response, StatusCode};
 use tokio::sync::Semaphore;
 use tokio::task::JoinSet;
 
-use crate::domain::nar::index::NarFileLocation;
-use crate::domain::nar::port::{NarStreamData, NarStreamHeaders, NarStreamProvider};
+use crate::domain::nar_info::index::NarFileLocation;
+use crate::domain::nar_info::port::{NarStreamData, NarStreamHeaders, NarStreamProvider};
 use crate::domain::substituter::model::Url;
 
 pub struct ReqwestNarStreamProvider {
